requirements.txt
Django==3.2
django-parler==2.3
Pillow==9.2.0
requests==2.28.1
beautifulsoup4==4.11.1
python-dotenv==0.21.0
postgres==4.0
psycopg2-binary==2.9.3
django-ckeditor==6.5.1
django-modeltranslation==0.18.4
